requirements.txt
psycopg2-binary>=2.9.9
pyodbc>=5.1.0
rich>=13.7.0
tabulate>=0.9.0
openai>=1.12.0
ollama>=0.4.0
groq>=0.9.0
sqlparse>=0.5.0
python-dotenv>=1.0.0
